Add `pybind11` and `Python` layers to `lightning.tensor` (#748)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**

Add both `python` and `pybind` layer to the `lightning.tensor`-`mps`
device.

1. Gates supported: Operators that have 1 or 2 wires or can be
decomposed into 1-2 wires gates. `MultiRZ` gate is not supported.
2. Obs supported: All obs supported in `lightning` state vector backends
except for `qml.Projector` and `SparseHamiltionian`
3. Both `qml.QubitStateVector` and `qml.StatePrep` are not supported.
4. `qml.state()` is not supported and the `state()` returns a
`LightningStateTensor` object represents the whole compute graph
5. `qml.BasisState` only accepts array representation of a basis state.
A `int` value can only supports up to `32` qubit system.
6. `qml.expval` is the only measurement `lightning.tensor` support.
7. Limited `gradient support` with `parameter-shift`. And a `100` qubits
SEL circuit was tested on `A100 80 GB GPU`
8. `maxBondDim` is default as `128`.
9. `pip` editable installation with `PL_BACKEND="lightning_tensor" pip
install -e .`, given that `cutensornet-cu12` is installed an `venv` env.
Detailed instructions are on[
PR#756](https://github.com/PennyLaneAI/pennylane-lightning/pull/756).
10. `lightning.tensor` is also tested against all the tests in `tests/`
if the tested feature is supported in `lightning.tensor`. The skipped
tests are tasks for the next step.
11. 

[SC-62440] [SC-62439] [SC-60329] [SC-60325]

**Description of the Change:**

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>
Co-authored-by: Mudit Pandey <mudit.pandey@xanadu.ai>
Co-authored-by: Vincent Michaud-Rioux <vincent.michaud-rioux@xanadu.ai>
Co-authored-by: Vincent Michaud-Rioux <vincentm@nanoacademic.com>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>