Fix the LK+Windows and Catalyst+Lightning latest/latest failures  (#758)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
- Remove `ControlledQubitUnitary` in favour of `C(QubitUnitary)` from
the list of supported operations and the device toml file. The
`stopping_condition` method guarantees the consistency of
decompositions. This change also reduces the size of the generated IR
and impact the performance of Catalyst for the computation of
`C(QubitUnitrary)` as this controlled gate is currently decomposed to
`QubitUnitary`, hardcoding the expanded controlled-unitary-matrix into
the IR.
- Raise a clear error message when initializing `lightning.kokkos` with
zero-qubit on Windows. Allocating 0-size Kokkos vectors not supported on
Windows. Required to Pass LK+Windows CI.
- Update toml files respectively to maintain support in Catalyst latest.
Required to Pass Catalyst-Lightning latest/latest CI.

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>