Update determine runner workflow to use large runners if the branch name matches a certain format (#769)
**Context:**
This PR is being opened under sc-66351. It introduces a new change to CI
ahead of release where a local branch (not forked) that matches the
regex pattern `vX.Y.Z-rcN` (The `N` after rc being optional) is auto
ported over to the large runners.

Though this may provide a minor bump to the runtimes of the workflows
themselves, the bigger objective (right now at least) is to not wait
long times to get a runner.

**Description of the Change:**
The change is a minor one within the `determine-runner` workflow. It now
checks for both a label present on a pr OR if the pr branch names
matches the rc branching format above.

**Benefits:**
Faster queue times for rc branches.

**Possible Drawbacks:**
Though the large runners and standard runners should have parity in
theory. We do not live in a perfect world. There is a none-zero chance
that workflows might break. But we can test that out as part of this PR.

**Related GitHub Issues:**
None.

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>