Optimize Lightning-Qubit's `probs(wires)` using bitshift implementation (#795)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
`probs` is central in circuit simulation measurements.

**Description of the Change:**
Implement `probs(wires)` using bitshift implementation akin to the gate
kernels in Lightning-Qubit. This does away with the NDpermuter
dependency, which is removed.

**Benefits:**
Enable on-the-fly diagonalization.
Faster execution.
The following benchmarks are performed on ISAIC's AMD EPYC-Milan
Processor using a single core/thread. The times are obtained averaging
the computation of `probs(target)` 5 times, where `target` includes a
single wire. The speed-ups are uniform across all possible targets, and
hence we show `target={0}` and the average speed-up across all targets
below.

![speedup_target0](https://github.com/user-attachments/assets/798a3385-d048-44e2-a0ce-4b34a8d7321b)

![speedup_targetavg](https://github.com/user-attachments/assets/afa26519-c1ca-4e04-92b5-508304ea0020)

The situation is more ambiguous when comparing w.r.t. the number of
targets. The following shows timings for `targets = {0..ntargets-1}` and
the conclusions are:
- The generic implementation is only faster for 1 target;
- The templated (specialized) implementation is faster for 1-4 targets,
which is a range where many `probs(Hermitian)` would be requested (since
at some point `Hermitian` becomes too large to represent as a matrix).


![speedup_vs_ntarget](https://github.com/user-attachments/assets/1ab5c1c6-4602-4126-a92f-f3c34d652a83)



**Possible Drawbacks:**

**Related GitHub Issues:**
[sc-65198]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>