Parallelize `probs` with OpenMP (#800)
### Before submitting

Please complete the following checklist when submitting a PR:

- [x] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [x] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [x] Ensure that the test suite passes, by running `make test`.

- [x] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [x] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:**
`probs` is central in circuit simulation measurements.

**Description of the Change:**
Parallelize `probs` loops using OpenMP.

**Benefits:**
Faster execution with several threads.
The following benchmarks are performed on ISAIC's AMD EPYC-Milan
Processor using a several core/threads. The times are obtained averaging
the computation of `probs(target)` 5 times for various number of
targets. We use the last release implementation as a reference. Since
#795 brings some speed-ups even for a single thread, this is why we
observe speed-ups > number of threads.


![speedup_vs_nthreads](https://github.com/user-attachments/assets/fdb762bb-8e50-4337-b2c0-7b16a42e8dad)

Another view on the data is the strong scaling efficiency. It is almost
perfect for 2-4 threads, fairly good for 8 threads and diminishes
significantly for 16 threads.


![efficiency_vs_nthreads](https://github.com/user-attachments/assets/4feca17e-a461-407c-947c-a0bc54c21b2a)

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>