TOML files update (#826)
**Context:** We would like Catalyst to optimize the first instance of
`qml.StatePrep` by just setting the state vector in PennyLane lightning.

**Description of the Change:** To achieve this, we set a flag in the
TOML files where it denotes whether the value for the flag
`skip_initial_state_prep` used during decomposition. This flag is
already in PennyLane and will skip the initial state preparation.

There are other ways to achieve this but this is the least invasive way
to do so. Some other alternatives that were discussed:

1. Adding StatePrep to the list of supported operations: this is not a
good idea as it would lead StatePrep to not be decomposed ever.
2. Creating a new operation and map StatePrep directly into this new
operation only in the context of Catalyst. This would lead to setting
this new `qml` Operation in the list of supported operations on the
device, even though it would never be seen except when using with
Catalyst.

The way this was achieved was not to create a new `qml` Operation, only
an JAXPR/MLIR operation and bind `StatePrep` to this JAXPR operation and
lower it through MLIR. So, no new `qml` operations.

**Benefits:** Optimization

**Possible Drawbacks:** More flags. No one likes flags.

**Related GitHub Issues:** Will be followed by this
https://github.com/PennyLaneAI/catalyst/pull/955 pull request in
Catalyst.

[sc-69069]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>