Fix install paths for cuquantum in CI and docs (#831)
### Before submitting

Please complete the following checklist when submitting a PR:

- [ ] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [ ] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [ ] Ensure that the test suite passes, by running `make test`.

- [ ] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [ ] Ensure that code is properly formatted by running `make format`. 

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.


------------------------------------------------------------------------------------------------------------

**Context:** Current install directions for cuquantum allow use of the
RPATH to the expected directory with Python, but requires explicit
targets when using a C++ build. The docs and paths searched for
cuquantum right now often have the `lib` dir implicit in the path, which
causes an incorrect path to be found by CMake, as it is a child class of
the root. This PR fixes those such paths, allowing the C++ builds to
find the correct custatevec libraries, ensuring a successful cmake
configuration step.

**Description of the Change:** As above.

**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>