Update Lightning packaging to follow the PEP 517/518/621/660 standards (#832)
**Context:** This PR updates the Lightning Python packaging to the
latest standard. This allows us to eliminate pip~=22.0 from
requirements*.txt, comply with up-to-date build practices, and fix the
inability to build Lightning in place on newer Python environments.

**Description of the Change:** Most of the building configuration was
moved to the ``pyproject.toml`` file. We kept the ``setup.py`` file
working with setuptools in part of the configuration pre-processing. A
script with some pre-pre-processing was added to configure the
``pyproject.toml`` file for each backend.

**Benefits:** Follow the new standard. Allows the use of more up-to-date
pip versions. There will be no impact on the final user, as the
installation of the wheels is unaffected by the change in packaging.

**Possible Drawbacks:** When building and installing Lightning backends
from scratch, we need to run the configuration script. If not the
installation will default to ``lightning.qubit``, or the last
configuration.

**Related GitHub Issues:**

[sc-65227]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Lee James O'Riordan <mlxd@users.noreply.github.com>
Co-authored-by: Vincent Michaud-Rioux <vincentm@nanoacademic.com>