Migration of LightningKokkos to the new device API (#810)
Please complete the following checklist when submitting a PR:

- [X] All new features must include a unit test.
If you've fixed a bug or added code that should be tested, add a test to
the
      [`tests`](../tests) directory!

- [X] All new functions and code must be clearly commented and
documented.
If you do make documentation changes, make sure that the docs build and
      render correctly by running `make docs`.

- [X] Ensure that the test suite passes, by running `make test`.

- [X] Add a new entry to the `.github/CHANGELOG.md` file, summarizing
the
      change, and including a link back to the PR.

- [X] Ensure that code is properly formatted by running `make format`.

When all the above are checked, delete everything above the dashed
line and fill in the pull request template.

------------------------------------------------------------------------------------------------------------

**Context:**
Following the design of *Lightning Qubit* uses the New Device API end to
end, we can now migrate `lightning.kokkos` device to the new device API.

**Description of the Change:**
Migration of Lightning Kokkos to integrate the New device API. The list
of main features changes are
- Add the `state_vector`, `measurements`, and `adjoint-jacobian` classes
for `lightning.kokkos`
- Add the `simulate`, `jacobian`, `simulate_and_jacobian`, `vjp`, and
`simulate_and_vjp` `methods to lighting.kokkos`
- Update unit/integration tests for the new device API to work with
`lightning.kokkos`
- Check the full support for sampling in full parity with Lightning
Qubit
- Replace the old device API for Lightning Kokkos.

**Benefits:**
Full integration of **Lightning kokkos** with the new device API.

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-59207], [sc-68825]

---------

Co-authored-by: ringo-but-quantum <github-ringo-but-quantum@xanadu.ai>
Co-authored-by: Amintor Dusko <87949283+AmintorDusko@users.noreply.github.com>
Co-authored-by: Vincent Michaud-Rioux <vincentm@nanoacademic.com>
Co-authored-by: Ali Asadi <10773383+maliasadi@users.noreply.github.com>